perfkit-requirements-alt.txt
# Copyright 2014 PerfKitBenchmarker Authors. All rights reserved.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#   http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
rsa==3.1.4
absl-py
jinja2>=2.7
setuptools>=40.3.0
colorlog[windows]==2.6.0
blinker>=1.3
futures>=3.0.3
PyYAML==5.3.1
pint>=0.7
numpy==1.16.4
contextlib2>=0.5.1
pywinrm
timeout-decorator
google-cloud-datastore
google-cloud-monitoring
#git+https://github.com/GoogleCloudPlatform/PerfKitBenchmarker#egg=perfkitbenchmarker
